Separate logging into separate logger.js and simplify server.js
